Monitoring de preguntas, impresión del árbol de decisión, y soporte de getNode(n) para todos los nodos del árbol.

+++ app/src/main/java/is/ecci/ucr/projectami/DecisionTree/ID3.java
@@ -13,7 +13,7 @@ public class ID3 {
 		_nodeCounter = 0;
 	}
 	
-	public double getNodeCounter() {
+	public int getNodeCounter() {
 		return _nodeCounter;
 	}
